added LOCK + REP RET
added the lock prefix.
currently functions as a delayed prefix router to normal assembler functions.
that is to say: there's currently no difference between using/not using a lock prefix.
however, at some point in the future it might actually do something.
impl has an accompanying set of instructions that are allowed to be used with it (some of which aren't implemented yet) - using anything else with a lock prefix is an assembly error.

added the REP(z|e|nz|ne) RET instruction format.
ret was already included (obviously), and rep ret isn't very common on modern arch, but it's included none the less because of popularity among assembly fanatics alone.
they're all translated into exactly the same byte code as a normal ret instruction.

+++ Computer.h
@@ -2873,7 +2873,7 @@ namespace CSX64
 			case 2:
 				// Intel's reference has this instruction modify CF unnecessarily - leaving those lines in but commenting them out
 
-				temp_u8 = AL();
+				temp_u8 = (u8)AL();
 				temp_b = CF();
 
 				//CF() = false;
@@ -2900,7 +2900,7 @@ namespace CSX64
 				return true;
 
 			case 3:
-				temp_u8 = AL();
+				temp_u8 = (u8)AL();
 				temp_b = CF();
 
 				CF() = false;
